requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
xgboost
joblib
python-multipart
python-dotenv
requests